Dealt with rl_scheduler and BaseDataset
